Move parts to components folder, add special thanks to gunnarx2
