+++ src/index.ts
@@ -1,6 +1,6 @@
 import { UserOptions } from 'embla-carousel'
-import EmblaCarouselReact from './EmblaCarouselReact'
-import useEmblaCarousel from './useEmblaCarousel'
+import EmblaCarouselReact from './components/EmblaCarouselReact'
+import useEmblaCarousel from './components/useEmblaCarousel'
 
 export default EmblaCarouselReact
 export { UserOptions, useEmblaCarousel }
